Enum urlencoded::UrlDecodingError

pub enum UrlDecodingError {
    BodyError(BodyError),
    EmptyQuery,
}

An error representing the two possible errors that can occur during URL decoding.

The first and probably most common one is for the query to be empty, and that goes for both body and url queries.

The second type of error that can occur is that something goes wrong when parsing the request body.

Variants
BodyError(BodyError)

An error parsing the request body

EmptyQuery

An empty query string, either in body or url query